requirements.txt
aniso8601==1.1.0
click==6.6
decorator==4.0.10
Flask==0.11.1
Flask-HTTPAuth==3.2.1
Flask-RESTful==0.3.5
Flask-SQLAlchemy==2.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
pbr==1.10.0
pkg-resources==0.0.0
PyHamcrest==1.9.0
python-dateutil==2.5.3
pytz==2016.7
six==1.10.0
SQLAlchemy==1.1.1
sqlalchemy-migrate==0.10.0
sqlparse==0.2.1
Tempita==0.5.2
Werkzeug==0.11.11
